Kafka/actuate-processor/requirements.txt
kafka-python==2.0.1
paho-mqtt==1.6.1